requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.6
django-cors-headers==4.7.0
django-filter==24.2
djangorestframework==3.15.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
notificationapi-python-server-sdk==2.0.0
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
